I got distracted and refactored the visuals... fuck
but at least it looks better now!! and has more info... soon..

changes:
I've clamped it to show a maximum number of neurons, this defaults to 20.
If a layer must be clamped then it will display the clamped number at the bottom of the layer
i have labelled the input, hidden and output layers.
added dummy training metrics to be implemented later
added a legend explaining the colours of things
added padding to top and bottom of screen (pain in my fucking ass)

realised the threading i added before didnt actually move it to a separate thread properly?? still looking into that

oh yeah and I also started finishing features I worked on prior to this new visual refactor forgetting to commit the bloody refactor so im smart arent I :)