Few more changes
Fixed the activation of the neurons, turns out I was doing it a bit wrong.
added a browse to path for save and load paths.
added progress bars and an elapsed timer though i think its borked.
inference, added a preview of the uploaded image and a slot next to it which will display the predicted letter.
Fixed the timer and progress bars. looking gooooood now
started adding better inference code, something that can preview. will probably fix then change this later